Changes to the examples to address all issues and this code now runs on all 3 simulators.
I have added VCS.log for VCS. if you want the other simulator logs, contact me.
This does also run on EDA playground as verified by a friendly colleague.

The file design.sv was used on EDA Playground, (contents are commented out, but you can uncomment and add, it works instead of the files in the HDL directory)